Breakthrough from rust/actix to html/js/css
