+++ src/lib.rs
@@ -2,3 +2,5 @@ pub mod network;
 mod parsing;
 pub use parsing::{Parser, Parsing};
 pub mod routing;
+
+pub mod ui;
